requirements.txt
flask==2.2.2
psycopg2-binary==2.9.1
pytest==7.2.1
python-dotenv==0.19.0
